+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [WIP]
+- API v4
+- Support for getting records via external ids
+
 ## [5.2.0]
 - Added support for illuminate collections v9
 - Added tests for php 8.1